requirements.txt
astroid==3.0.2
blinker==1.4
cffi==1.16.0
click==8.1.7
coverage==7.4.1
cryptography==3.4.8
dbus-python==1.2.18
dill==0.3.8
distlib==0.3.8
distro==1.7.0
exceptiongroup==1.2.0
Faker==25.8.0
filelock==3.14.0
httplib2==0.20.2
importlib-metadata==4.6.4
iniconfig==2.0.0
isort==5.13.2
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
mccabe==0.7.0
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.0.0
oauthlib==3.2.0
packaging==23.2
pandas==2.2.2
platformdirs==4.1.0
pluggy==1.4.0
pycairo==1.26.0
pycparser==2.22
PyGObject==3.42.1
PyJWT==2.3.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pylint-pytest==1.1.7
pyparsing==2.4.7
pytest==8.0.0
pytest-cov==4.1.0
pytest-pylint==0.21.0
pytest-timeout==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
SecretStorage==3.3.3
six==1.16.0
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
wadllib==1.3.6
zipp==3.19.2
